requirements.txt
## The following requirements were added by pip freeze:
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.4
openai==0.23.1
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.221012
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
rsa==4.9
scipy==1.9.2
simpleaudio==1.0.4
six==1.16.0
sounddevice==0.4.5
soundfile==0.11.0
tqdm==4.64.1
types-pytz==2022.4.0.0
typing_extensions==4.4.0
urllib3==1.26.12
